+++ CHANGELOG.md
@@ -37,6 +37,7 @@ release.
 ### Changed
 - Enabled Hayabusa2 drivers [#596](https://github.com/DOI-USGS/ale/pull/596)
 - Enabled Juno drivers [#597](https://github.com/DOI-USGS/ale/pull/597)
+- Enabled Odyssey drivers [#582](https://github.com/DOI-USGS/ale/pull/582)
 
 ### Added
 - Apollo Metric drivers, tests, and data [#533](https://github.com/DOI-USGS/ale/pull/533)